first real QT4 port changes
git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/branches/QT4@1508 6adb6e08-d915-0410-941f-83917bcadc18

+++ include/OpenMS/APPLICATIONS/TOPPViewBase.h
@@ -40,7 +40,6 @@
 //QT
 #include <q3mainwindow.h>
 #include <qworkspace.h>
-//Added by qt3to4:
 #include <QCloseEvent>
 #include <QLabel>
 #include <Q3PopupMenu>
@@ -138,7 +137,7 @@ namespace OpenMS
         maximizeActiveSpectrum();
         for (int i=0; i<tab_bar_->count(); ++i)
         {
-        	if (tab_bar_->tabData(i).toInt() == int(ws_->activeWindow()))
+        	if (tab_bar_->tabData(i).toInt() == qlonglong(ws_->activeWindow()))
         	{
         		tab_bar_->setCurrentIndex(i);
         		break;
